[sqm-scripts] Make the makefile work with BSD/OSX utilities as well
The Makefile allows to create nice folders containig the whole set of
acripts for a plattform. Unfortunately the macosx install binary
does not know about the -t switch, the workaround seems to be simply
putting the destination directory at the end of the list of files
to copy. This change does axactly that; tested to work under macosx 10.9.5
and opensuse 13.2.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>