[sqm-scripts] Introduce a new qos script for cake
Unitil now we introduced cake as an option into existing
qos scripts, but since cake acts both as shaper as well as leaf
qdisc this resulted in code less clear than wished for. This script
will hopefully allow to remove the cake as replacement for htb switch
statement from the other scripts.
Signed-off-by: Sebastian Moeller <moeller0@gmx.de>