[smq-scripts] Teach run-openwrt.sh to only stop actually runnning sqm instances
The remodelled run.sh was trying to stop itself for all configured
sections in /etc/config/sqm, including sections that were neither
enabled nor had a running instance of sqm. Now run.sh will only try to sqm-stop
running instances. This was not incuded in stop-sqm, so that stop-sqm
will still work in case the system to remember the running state gets confused...
This fix gets rid of a number of unnecessary error messages...

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>