[sqm-scripts] Use only verified qdiscs
Sqm-scripts and luci-app-sqm just learnd to verify which of a list
of qdiscs specified in run.sh can actually be attached to an ifb
without error, in other words which qdiscs from the list should be useable.
Now we actually go and use only those. ATTENTION this requires a recent
matching luci-app-sqm to actually work. Needs testing and Toke's buy-in.
Otherwise we need to come up with a backward compatible way of switching
over from the old style.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>