Restructure the qdisc verification code.
This restructures the code verifying the existence of qdiscs. The
verify_qdisc function now uses return values to signal success/error, as
does a couple of supporting functions. It also supports passing a list
of supported qdiscs as the second parameter.

The list of available qdiscs is now in a directory specified in the
config file. The update-available-qdiscs script will create a file named
after each qdisc it finds there. On openwrt, this script is run at
application time if the directory doesn't exist. The list of supported
qdiscs is now built by explicitly testing for a list of pre-defined
qdiscs, as well as by looking at the available kernel modules.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>