Check availability of ongress qdisc during setup of ingress shaper
If the ingress qdisc is not available and useable sqm-scripts will
not behave as the user expects. So this commit teaches verify_qdisc
the peculiarities of the invocatin of sch_ingress and switches all
actual qos scripts to verify ingress unless ingress shaping was
deselected by the user (download bandwidth set to zero).