Fix the cut here marker to keep logger happy.
The new diafnostic output of get_ifb_associated_with_if included
a line denoting from where to paste data into an error report.
Unfortunatelly this line started with - which logger misttok as
an indication of a coming option, that was not there. This replaces
the first (and last) - with a #, and now it actually works.