Teach sqm_logger to also pipe its output into a debug log file
To facilitate remote debugging sessions with users it is convenient
to have a concise output of sqm scripts log messages without
the risk for the user of sharing irrelvant info from their log.
To allow this sqm_logger will also echo a copy of its output strings
into the file specified by SQM_DEBUG_OUTPUT_FQN in sqm.conf. As long
as the path exists...
The current behaviour is somewhat dangerous as sqm will continue to add
entries to this file instead of deleting it with each "/etc/init.d.sqm start"
this behaviour might need changing if deemed to risky. SQM_DEBUG_OUTPUT_FQN
is empty by default, so it will require user intervention to become an
issue.