Make the automatic target increase also do the right thing for ATM links
sqm-scripts tries to scale codel's target variable so that one
full sizes on-the-wire packet can be transferred without exceeding
the target time. Up till now this was based on a hypothetical 1540
byte packet (including som overhead). Wich seems to work reasonably well
but it does fail to account for ATM links celling; on an ATM link a
packet with roughly MSS 1460 get extended into 32 or 33 cells of 53 bytes.
To account for that target now is scaled not for one packet of 1540 bytes
but of 33 * 53 = 1749 bytes to make sure this also works for ATM links.
In reality it seems the target autoscaling only needs to be approximately
correct so hopefully this change should not lead to large behavioral
changes (needs confirmation, before the next change, scaling target to to
one packet but to 1.5...).

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>