Start conversion to use log levels for sqm_logger
Introduce the concept of graded loging into start-sqm and stop-sqm.
To reduce sqm's verbosity all calls to sqm_logger need to be revisited
regarding their importance. Defaults.sh introduces symbolic names for the
current 4 log levels, _SILENT, _NORMAL, _VERBOSE, and _TOO_MUCH, all
prefixed with VERBOSITY. start-sqm and stop-sqm are changed to use the new
constructs for non-essential messages. Also this commit reduces the
number of lines of a successful call to start-sqm from 2 to 1, adds
the output of date to both start-sqm and stop-sqm logging output
to increase the value of debug logging files. Also the message of stop-sqm
has been changed closer to start-sqm's wording. Passes preliminary testing...

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>