Remove date putput from sqm_logger calls
To avoid redundant information in sqm syslog output, only use
date to create a time marker in the debug log every time eithrt
start-sqm or stop-sqm was running. Since the debug log is not
deleted by default this information should allow to make more
sense out of the debug logs contents in case of multiple sqm
cycles logged to the same file.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>