Silence errors when running during boot.
Oftentimes, the configured interface will not be available during boot,
and so we will output an error message. This is fine, as the script will
be re-run later by hotplug, but it gives misleading error messages in
the boot logs. However, we can't get rid of the start-at-boot entirely,
since there can be cases where no hotplug event will be sent for an
interface (e.g. bridge interfaces). So instead, we just silence errors
during boot.

This adds an SQM_VERBOSITY_MIN setting (and renames the old
SQM_VERBOSITY to SQM_VERBOSITY_MAX), and sets it during boot in the init
script.

Closes #46.