tools/sqm.lua: Provide common Luci SQM helper utilities
This factors out and defines various functions for reading and parsing
QDISC capabilities, and some specialized GUI option handling, to be used
by the other Luci SQM pages.

Define read_caps() to read qdisc and shaper capabilities, and cache them
within a GUI page load. This uses the helper script 'get-qdisc-caps' to
retrieve and parse a JSON QDISC capabilities string associated with the
sections 'script' option. The function attempts to get 'script' using
formvalue() first before cfgvalue(). This allows a user to press only
"Save" after selecting a custom script on the 'details' page to reload
the page and have related QDISC changes reflected in the GUI.

Define the match_caps_pairs() function, which uses a capability filter to
iterate over a qdisc list or capability list.

Define the function parse_tuple_caps(), which extracts details of 3-tuple
capabilities, including type, variable values and related descriptive text.

Factor out code for AbstractSection.varianttaboption(), used to set
up "variant" options that implement complex dependencies within the GUI,
allowing for varying the displayed option values based on carefully
defined dependencies.

Provide an AbstractValue.write() replacement that conditionally enables
the related init script. This is used on the advice of Jo-Philipp Wich
and following the example of the luci-app-minidlna CBI model, to help
users by enabling sqm's init script if even a single sqm instance is
enabled and state is saved ("Save & Apply" or "Save").
This version additionally checks that the init script is not already
enabled, to avoid spamming the GUI with notifications.