Added cautious handling of mpu handling for cake
We have multiple versions of cake in the field with different
sets of available parameters, like the minimal packet size to
use for overhead and link layer accounting "mpu". This change will
make cake use the mpu parameter (tcMPU) if the used tc binary
will report that cake supports the mpu parameter via the tc help
output (tc qdisc add root cake help). In that case sqm will attempt
to pass "mpu $tcMPU" to cake, if cake is selected as qdisc.
This introduces a helper function verify_qdisc_parameter_against_tc_help
that will take the qdisc as $1 and the parameter as $2 and returns
TRUE and FALSE (strings).

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>