functions.sh: Minimize using temporary QDISC/IFB in verify_qdisc()
Rewrite the funtion to use a series of checks in order from most
preferable to least. These begin with various passive module
introspection methods and end by instantiating the qdisc on a temporary
IFB device as a last resort.

This last, heavy-weight approach should be avoided as much as possible,
so in addition we cache results, noting it should only be required for
builtin qdisc modules on minimal (or broken) Linux distros.

These changes help reduce undesirable network impacts and permit using
verify_qdisc() with fewer caveats.