Fix an embarrasing typo in do_modules
As guidosarducci recognized the change that intended to
avoid loading already loaded modules again and thereby
avoiding an failure reports reaching the log contained
a typo and hence did always unconditionally tried to
load modules and caused log spam.
Fixing this typo hopefully keeps the log a bit tidier.
And as guidosarducci commented in
https://github.com/guidosarducci/sqm-scripts/commit/00cd65dc527913aa75fb15e00c016212dae02828
"This has been broken since the original commit 77744c1 2 years ago, for
want of simply typing the code into a running shell as a basic check."
Thanks for both the fix and the snark.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>