Base the  burst buffer size for htb & tbf on the time required to empty at the spcified bandwidth.
This should make it more obvious how much latency under load increase
the htb's burst-buffer might cause. Ideally we would not need to
set this manually, but the default size of 1600B will only work well
for low bandwiidth links, for faster links the buffer will empty
before the shaper is executed again and lost transmit opportunities
will show as lost bandwidth. Especially at with internet access
links now reaching into the multiple 100s of Mbps the existing
automatic buffer scaling proofed to be sub-optimal.
The old functions tried to scale the buffer in stpes of full
MTUs but accidentally also sized the buffer proportional to
the defined MTU, and also stopped scaling above 216 Mbps
(and due to a bug that went unnoticed for too long instead of
keeping the maximum burst buffer of 15KB for babdwidth >= 216 Mbps
it fell back to the default 1600B, massively affecting achievable
bandwidth).)

Thanks to github users tohojo, dtaht negan07, and guidosarducci
for discussion and feedback!

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>