Reire the "classic" htb burst calculation method
Instesd of usinf the boolean varable SHAPER_BURST to
use the implicit bust calculation method in sqm-scripts
exclusively switch to using TARGET_BURST_DUR_MS ( unless over-ridden
by the directional ITARGET_BURST_DUR_MS and ETARGET_BURST_DUR_MS)
to configure the size of HTB's burst buffer. This method will scale
the buffer to be "cleared" in the specified number of milliseconds.
If the duration is set to zero it will use HTB's default value instead.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>