Switch HTB's quantum calculations to be duration based
This homogenizes the burst and quantum calculations for HTB
Quantum defaults to be equal to Burst. Also some small fixes
Switch HTB's quantum calculations to be duration based

This homogenizes the burst and quantum calculations for HTB
Quantum defaults to be equal to Burst. Also some small fixes
to the burst by duration functions.to the burst by duration functions.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>