Trigger the sqm_error output on LAST_ERROR not being empty instead of on the exit status
It seeems that exist status passing from sub-shell invocations
to the caller are not as robust as desired.
Since Sucessful calls to ip and tc do not create standard error output
we simply log any non-empty outputs of the sub-shell as errors.
That seems to work and avoids duplicate calls og the same
tc/ip invocation. This affects both ip_wrapper and tc_wrapper.
Passes testing on OpenWrt master r8311.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>