Simplyfy fn_exists() while keeping compatibility
This tries to get rid of the special casing of Debian's
dash? type builtin. It also fixes a typo in the ouput and
evaluates type only once and greps the saved output string.
This also removes the saveing on LC_ALL since with the
sub shell execution the caller's LC_ALL should be left alone
anyway.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>