Logging changes
Instead of using SQM_DEGUG to append all output into
a growing logfile, use three logfiles instead.
One for common code, mostly from run.sh, one for start-sqm and one for stop-sqm.
These logs are overwritten respectively on each invocation of either
run.sh, start-sqm or stop-sqm and will not grow indefintely and hence
can be enabled unconditionally so that this information is
available for debugging directly from "broken" sqm instances.

This commit also introduces ipt_new() in an attempt to convert
the legacy ipt() function to use cmd_wrapper to actually call
the iptables/ip6tables binaries. This exposes loads of errors
in ip[6]tables output and is hence not activated as default yet.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>