Silence the cutionary attempts to delete iptables cahins before addition or insertion
ipt() seems to first try to delete chains it wants to add or insert to
Given that we run sqm_stop explicitly, this seems not really necessary
Here we just use SILENT=1 to suppress the erros from trying to delete
non-existing chains. This should not be any worse than the old code
but since the errors are still unconditionally logged, should be
considered a step in the right direction..Here we just use SILENT=1 to suppress the erros from trying to delete
non-existing chains. This should not be any worse than the old code
but since the errors are still unconditionally logged, should be
considered a step in the right direction....

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>