Switch ipt() from custom logging to employing cmd_wrapper()
The recent introduction of cmd_wrapper for important binaries
(tc, ip) left iptables and ip6tables usage in ipt() to use
custom logging. This switches ipt() over to use the wrappers,
To avoid log spew and to keep the situation compatible to the
status quo, this will sielnce all calls to ip[6]tables in ipt().

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>