Improve logging for SILENT=1 invocations
If a cmd_wrapper invocation is performed with SILENT set to 1
cmd_wrapper will demote sqm_error to sqm_debug to keep the
log/syslog clean. Now these Errors still make it into the degug_log.
Here we simply add a line indicating that failures for the next command
are acceptable. This also improves the comment on the SILENT variable in
defaults.sh.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>