Modify sqm's logging to allow enable debug logging by default (#99)
* Modify sqm's logging to allow enable debug logging by default

Instead of the old continously growing log approach, this switches to
one log file each for start-sqm and stop-sqm. This allows to truncate
each log at the start of each of the two scripts, so that effectively
for each interface with an sqm instance there will be a log for the
most recent start-sqm and stop-sqm instance. Unlike the old growing log
this does not risk to fill up /varrun with its verbosity, and
more importantly allows to enable SQM_DEBUG by default to that for
every issue encountered in the field we will be able to get to
a log file from eactly the failed instance (unless users manually
disable this).

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>
[ squash into one commit ]
Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>