Makefile: Drop creation of state dir
Since /run is often a temporary filesystem, there's not much benefit to
creating /run/sqm at 'make install' time. Debian's lintian seems to think
it is even an error to do so.

We're already checking for the run directory in functions.sh, and creating
it if it doesn't exist, so this should have no impact on behaviour.

Fixes #104.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>