Switch default configs to using piece_of_cake.qos
Since Cake has now been in upstream Linux for a while, let's start using it
as the default. For OpenWrt we already depend on kmod-sched-cake, so switch
the default here as well.

Also use the backup facility of 'install' when installing the config
files (default.confg and sqm.conf), to make sure we don't clobber a user's
config on 'make install'.

Fixes #112.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>