linux/sqm-bin: Source config files in subshell to prevent vars leaking
When starting sqm-scripts on all interfaces, the config files were sourced
in the calling shell instead of the subshell. This led to leaking of config
variables that were defined in one file but not in others. Fix this by
moving the config sourcing into the subshell (as was already done on the
'stop' action).

Fixes #113.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>