functions: Simplify write_state_file
There is really no reason to have a wrapper function to extract the names
when we're always using it the same way. So let's get rid of the long
function name for the wrapper function, and just parse defaults.sh directly
in write_state_file.

While we're at it, use a single awk script instead of a grep/sed combo to
extract the variables, and purge any duplicates.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>