defaults: Look for iptables-nft if iptables binary not found
For systems switching to nftables, the iptables-nft binary can be used as a
compatibility layer for applications relying on iptables. Not every
distribution (notably OpenWrt) installs a symlink from iptables-nft to
iptables, though, so add a fallback check for iptables-nft if we can't find
iptables itself.

Both binaries can be present on the same system simultaneously, so we can't
really infer anything from their presence. In the interest of not breaking
existing setups, prefer the regular 'iptables' binary over 'iptables-nft'.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>