openwrt: Default to iptables-nft
OpenWrt is moving to using nftables by default, but the 'iptables' binary
is likely to stick around for a while. So add a platform-specific
preference for using iptables-nft (if it exists). We'll still fall back to
'iptables' if 'iptables-nft' is not installed on the system, and we'll
still get the warning if neither exists and simple.qos is used.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>