functions.sh: Split out cleanup into separate sqm_cleanup function
Split out the cleaning of iptables rules and IFB interface to a separate
sqm_cleanup() function, which can optionally be called with a "silent"
parameter which will silence the IFB removal command. This is in
preparation for exposing this as a separate cleanup action that can be
called after the main netdevice is removed.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>